python/src/auth/requirements.txt
PyJWT==2.6.0
pylint==2.15.5
Jinja2==3.1.2
Flask==2.2.2
cryptography==38.0.3
psycopg2-binary==2.9.5